requirements.txt
absl-py==0.9.0
aenum==2.2.6
astroid==2.4.1
astunparse==1.6.3
attrs==20.3.0
autopep8==1.5.3
beautifulsoup4==4.9.3
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
cycler==0.10.0
fbchat==1.9.7
gast==0.3.3
google-auth==1.16.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
idna==2.9
isort==4.3.21
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
Markdown==3.2.2
matplotlib==3.2.1
mccabe==0.6.1
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.2.1
paho-mqtt==1.5.1
pandas==1.0.4
Pillow==7.1.2
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pylint==2.5.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.15.0
soupsieve==2.1
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
toml==0.10.1
urllib3==1.25.9
Werkzeug==1.0.1
wordcloud==1.7.0
wrapt==1.12.1
